Fixed exceptions in `_trigger()`
Now correctly throws exception for all scenarios:

If the transition was aborted, a `CouldNotTransitionExceptionInterface` is *always* thrown. If an exception is thrown by the event handlers , it will be attached as an inner exception.
If the transition was not aborted and an exception was thrown by an event handler, the transition will be committed and a `StateMachineExceptionInterface` is thrown with the event handler's exception attached as an inner exception.